move functions to routerUtils
