+++ util/customTransforrmationsStore-v1.js
@@ -7,7 +7,7 @@ const libraryCache = {};
 
 // const CONFIG_BACKEND_URL =
 //   process.env.CONFIG_BACKEND_URL || "https://api.rudderlabs.com";
-const CONFIG_BACKEND_URL = "https://api.dev.rudderlabs.com";
+const CONFIG_BACKEND_URL = "http://localhost:5000";
 const getTransformationURL = `${CONFIG_BACKEND_URL}/transformation/getByVersionId`;
 const getLibrariesUrl = `${CONFIG_BACKEND_URL}/transformationLibrary/getByVersionId`;
 
@@ -47,7 +47,9 @@ async function getLibraryCodeV1(versionId) {
   };
   try {
     const startTime = new Date();
-    const response = await fetchWithProxy(`${getLibrariesUrl}?versionId=${versionId}`);
+    const response = await fetchWithProxy(
+      `${getLibrariesUrl}?versionId=${versionId}`
+    );
     stats.increment("get_libraries_code.success", tags);
     stats.timing("get_libraries_code", startTime, tags);
     const myJson = await response.json();
