+++ util/customTransforrmationsStore.js
@@ -1,5 +1,5 @@
-const { fetchWithProxy } = require("./fetch");
 const NodeCache = require("node-cache");
+const { fetchWithProxy } = require("./fetch");
 const logger = require("../logger");
 const stats = require("./stats");
 
@@ -7,7 +7,7 @@ const myCache = new NodeCache();
 
 // const CONFIG_BACKEND_URL =
 //   process.env.CONFIG_BACKEND_URL || "https://api.rudderlabs.com";
-const CONFIG_BACKEND_URL = "https://api.dev.rudderlabs.com";
+const CONFIG_BACKEND_URL = "http://localhost:5000";
 const getTransformationURL = `${CONFIG_BACKEND_URL}/transformation/getByVersionId`;
 
 // Gets the transformation from config backend.
