+++ versionedRouter.js
@@ -4,12 +4,19 @@
 const heapdump = require("heapdump");
 const Router = require("koa-router");
 const _ = require("lodash");
-const { lstatSync, readdirSync } = require("fs");
 const fs = require("fs");
 const logger = require("./logger");
 const stats = require("./util/stats");
 const { isNonFuncObject, getMetadata } = require("./v0/util");
-const { DestHandlerMap } = require("./constants/destinationCanonicalNames");
+const {
+  getIntegrations,
+  getDestHandler,
+  getDestFileUploadHandler,
+  getPollStatusHandler,
+  getJobStatusHandler,
+  handleDestinationNetwork,
+  userTransformHandler
+} = require("./routerUtils");
 require("dotenv").config();
 
 const versions = ["v0"];
@@ -24,46 +31,6 @@ const networkMode = process.env.TRANSFORMER_NETWORK_MODE || true;
 
 const router = new Router();
 
-const isDirectory = source => {
-  return lstatSync(source).isDirectory();
-};
-
-const getIntegrations = type =>
-  readdirSync(type).filter(destName => isDirectory(`${type}/${destName}`));
-
-const getDestHandler = (version, dest) => {
-  if (DestHandlerMap.hasOwnProperty(dest)) {
-    return require(`./${version}/destinations/${DestHandlerMap[dest]}/transform`);
-  }
-  return require(`./${version}/destinations/${dest}/transform`);
-};
-
-const getDestNetHander = (version, dest) => {
-  const destination = _.toLower(dest);
-  let destNetHandler;
-  try {
-    destNetHandler = require(`./${version}/destinations/${destination}/nethandler`);
-    if (!destNetHandler && !destNetHandler.sendData) {
-      destNetHandler = require("./adapters/networkhandler/genericnethandler");
-    }
-  } catch (err) {
-    destNetHandler = require("./adapters/networkhandler/genericnethandler");
-  }
-  return destNetHandler;
-};
-
-const getDestFileUploadHandler = (version, dest) => {
-  return require(`./${version}/destinations/${dest}/fileUpload`);
-};
-
-const getPollStatusHandler = (version, dest) => {
-  return require(`./${version}/destinations/${dest}/poll`);
-};
-
-const getJobStatusHandler = (version, dest) => {
-  return require(`./${version}/destinations/${dest}/fetchJobStatus`);
-};
-
 const eventValidator = require("./util/eventValidation");
 const getSourceHandler = (version, source) => {
   return require(`./${version}/sources/${source}/transform`);
@@ -75,13 +42,6 @@ const functionsEnabled = () => {
     areFunctionsEnabled = process.env.ENABLE_FUNCTIONS === "false" ? 0 : 1;
   }
   return areFunctionsEnabled === 1;
-};
-
-const userTransformHandler = () => {
-  if (functionsEnabled()) {
-    return require("./util/customTransformer").userTransformHandler;
-  }
-  throw new Error("Functions are not enabled");
 };
 
 async function handleDest(ctx, version, destination) {
@@ -518,34 +478,6 @@ if (startSourceTransformer) {
       });
     });
   });
-}
-
-async function handleDestinationNetwork(version, destination, ctx) {
-  const destNetHandler = getDestNetHander(version, destination);
-  // flow should never reach the below (if) its a desperate fall-back
-  if (!destNetHandler || !destNetHandler.sendData) {
-    ctx.status = 404;
-    ctx.body = `${destination} doesn't support transformer proxy`;
-    return ctx.body;
-  }
-  let response;
-  logger.info("Request recieved for destination", destination);
-  try {
-    response = await destNetHandler.sendData(ctx.request.body);
-  } catch (err) {
-    response = {
-      status: 500, // keeping retryable default
-      error: err.message || "Error occurred while processing payload."
-    };
-    // error from network failure should directly parsable as response
-    if (err.networkFailure) {
-      response = { ...err };
-    }
-  }
-
-  ctx.body = { output: response };
-  ctx.status = response.status;
-  return ctx.body;
 }
 
 if (networkMode) {
